+++ splash/tests/test_execute_element.py
@@ -1,11 +1,13 @@
 # -*- coding: utf-8 -*-
 import base64
 from io import BytesIO
+
 from PIL import Image
 import pytest
 
 lupa = pytest.importorskip("lupa")
 
+from splash import defaults
 from splash.exceptions import ScriptError
 from .test_execute import BaseLuaRenderTest
 
@@ -109,8 +111,11 @@ class HTMLElementTest(BaseLuaRenderTest):
             end
 
             assert(body:mouse_click(0, 0))
+            assert(body:mouse_click{x=0, y=0})
             assert(body:mouse_click(5, 10))
+            assert(body:mouse_click{x=5, y=10})
             assert(body:mouse_click(20, 40))
+            assert(body:mouse_click())
 
             assert(splash:wait(0))
 
@@ -119,9 +124,38 @@ class HTMLElementTest(BaseLuaRenderTest):
         """, {"url": self.mockurl("various-elements")})
 
         self.assertStatusCode(resp, 200)
+
+        w, h = map(int, defaults.VIEWPORT_SIZE.split('x'))
         self.assertEqual(resp.json(), [
-            {'x': 0, 'y': 0}, {'x': 5, 'y': 10}, {'x': 20, 'y': 40}
+            {'x': 0, 'y': 0},
+            {'x': 0, 'y': 0},
+            {'x': 5, 'y': 10},
+            {'x': 5, 'y': 10},
+            {'x': 20, 'y': 40},
+            {'x': w//2, 'y': h//2}
         ])
+
+    def test_mouse_click_border_radius(self):
+        HTML = """
+        <html>
+        <body>
+            <a style="padding:10px;border-radius:7px"
+               onclick="document.write('clicked');">
+                click me
+            </div>
+        </body>
+        </html>        
+        """
+        resp = self.request_lua("""
+        function main(splash)
+            assert(splash:set_content(splash.args.html))
+            assert(splash:select('a'):mouse_click())
+            assert(splash:wait(0))
+            return splash:html()
+        end
+        """, {'html': HTML})
+        self.assertStatusCode(resp, 200)
+        self.assertEqual(resp.text.count("onclick"), 0, resp.text)
 
     def test_mouse_click_bad_x_argument(self):
         resp = self.request_lua("""